Move `ChatView` to seperate folder
