+++ lib/base/networking/api/handler/settings_handler.dart
@@ -92,7 +92,7 @@ class SettingsHandler {
 
   /// Updates the preferred greeting in user settings.
   Future<bool> updatePreferredGreeting(
-      String newGreeting, List<UserSetting> currentSettings) async {
+      String newGreeting, List<UserSetting> currentSettings,) async {
     try {
       var greetingSetting = currentSettings.firstWhere(
         (setting) => setting.type == UserSettingType.GREETING,
@@ -114,7 +114,7 @@ class SettingsHandler {
 
   /// Updates the preferred name in user settings.
   Future<bool> updatePreferredName(
-      String newName, List<UserSetting> currentSettings) async {
+      String newName, List<UserSetting> currentSettings,) async {
     try {
       var newSetting =
           UserSetting(type: UserSettingType.PREFERRED_NAME, value: newName);
@@ -128,7 +128,7 @@ class SettingsHandler {
 
   /// Updates the selected speeds in user settings.
   Future<void> updateSelectedSpeeds(
-      double speed, bool isSelected, List<UserSetting> currentSettings) async {
+      double speed, bool isSelected, List<UserSetting> currentSettings,) async {
     var playbackSpeedSetting = currentSettings.firstWhere(
       (setting) => setting.type == UserSettingType.CUSTOM_PLAYBACK_SPEEDS,
       orElse: () => UserSetting(
